pyproject.toml
[project]
name = "egm-horeca-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kadyr Gullyyev",email = "kadyr.gullyyew@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
